Fix Refactor Converter
